requirements.txt
azure-identity
azure-storage-blob
azure-storage-file-datalake